unreleased Async-0.15

+++ Async.pm
@@ -1,19 +1,23 @@
 
 package Async;
-$VERSION = '0.10';
+$VERSION = '0.15';
 
 sub new {
   my ($pack, $task) = @_;
+  unless (defined $task) {
+    require Carp;
+    Carp::confess("Usage: Async->new(sub { ... })");
+  }
   my $r   = \ do {local *FH};
   my $w = \ do {local *FH};
   unless (pipe $r, $w) {
-    $ERROR = "Couldn't make pipe: $!";
-    return;
+    return bless { ERROR => "Couldn't make pipe: $!",
+		   FINISHED => 1} => $pack;
   }
   my $pid = fork();
   unless (defined $pid) {
-    $ERROR = "Couldn't fork: $!";
-    return;
+    return bless { ERROR => "Couldn't fork: $!",
+		   FINISHED => 1} => $pack;
   }
   if ($pid) {			# parent
     close $w;
@@ -33,7 +37,7 @@ sub new {
 }
 
 # return true iff async process is complete
-# with true `$force' argmuent, wait until process is complete before returning
+# with true `$force' argument, wait until process is complete before returning
 sub ready {
   my ($self, $force) = @_;
   my $timeout;
@@ -43,7 +47,7 @@ sub ready {
   vec($fdset, $self->{FD}, 1) = 1;
   while (select($fdset, undef, undef, $timeout)) {
     my $buf;
-    my $nr = read $self->{PIPE}, $buf, 8192;
+    my $nr = sysread $self->{PIPE}, $buf, 8192;
     if ($nr) {
       $self->{DATA} .= $buf;
     } elsif (defined $nr) {		# EOF
@@ -82,9 +86,64 @@ sub result {
 sub DESTROY {
   my ($self) = @_;
   my $pid = $self->{PID};
+  return unless defined $pid;
   kill 9 => $pid;	# I don't care.
   waitpid($pid, 0);
 }
+
+
+################################################################
+
+sub TIEHASH {
+  my $pack = shift;
+  my %self;
+  bless \%self => $pack;
+}
+
+sub STORE {
+  my ($self, $key, $val) = @_;
+  $self->{$key} = (ref $self)->new($val);
+}
+
+sub FETCH {
+  my ($self, $key) = @_;
+  my $type = ref $self->{$key};
+  if ($type) {
+    if ($self->{$key}->ready) {
+      my $result = $self->{$key}->result;
+      $self->{$key} = $result;
+      return $result;
+    } else {
+      return;
+    }
+  } else {
+    return $self->{$key};
+  }
+}
+
+sub EXISTS {
+  my ($self, $key) = @_;
+  my $type = ref $self->{$key};
+  return $type ? $self->{$key}->ready : $self->{$key};
+}
+
+sub DELETE {
+  my ($self, $key) = @_;
+  delete $self->{$key};
+}
+
+sub FIRSTKEY {
+  my ($self) = @_;
+  keys %$self;
+  return scalar(each %$self);
+}
+
+sub NEXTKEY {
+  my ($self) = @_;
+  return scalar(each %$self);
+}
+
+################################################################
 
 package AsyncTimeout;
 @ISA = 'Async';
@@ -100,10 +159,21 @@ sub new {
       return $msg if !defined($s) && $@ eq "TIMEOUT\n";
       return $s;
     };
+  
+  # A timeout of 0 doesn't work as you might expect, because 0 has a
+  # special meaning to the alarm() call.  So if the timeout is zero,
+  # we'll use a fake computation that behaves *as if* it had timed out
+  # immediately.
+  $newtask = sub { $msg } if $timeout == 0;  # I like this hack.
+
   my $self = Async->new($newtask);
   return unless $self;
   bless $self => AsyncTimeout;
 }
+
+################################################################
+
+
 
 package AsyncData;
 @ISA = 'Async';
@@ -114,7 +184,7 @@ sub new {
   my $newtask =
     sub {
       my $v = $task->();
-      return Storable::freeze($v);
+      return ref $v ? Storable::freeze($v) : $v;
     };
   my $self = Async->new($newtask);
   return unless $self;
@@ -125,7 +195,66 @@ sub result {
   require Storable;
   my $self = shift;
   my $rc = $self->SUPER::result(@_);
-  return defined $rc ? Storable::thaw($rc) : $rc;
+  return unless defined $rc;
+  my $result =  Storable::thaw($rc);
+#  print STDERR "rc: ($rc) result: ($result)\n";
+  return defined $result ? $result : $rc;
+}
+
+
+################################################################
+
+
+
+package AsyncCallback;
+@ISA = 'Async';
+
+my @callbacks;
+my @args;
+
+sub _run_callbacks {
+  local $_;
+  for (@callbacks) { next unless defined $_;
+                     my ($f, $v) = @$_; 
+                     $f->($v);      # $_->[0]->($_->[1])?
+                   } 
+  $oldHandler->('USR1') if ref $oldHandler ;
+}
+
+sub new {
+  my ($pack, $task, $callback, $arg) = @_;
+  my $n;
+  if ($callback) {
+    unless (@callbacks) {
+      $oldHandler = $SIG{USR1} || 'DEFAULT';
+      $SIG{USR1} = \&_run_callbacks;
+    }
+    push @callbacks, [$callback, $arg];
+    $n = $#callbacks;
+  }
+  my $pid = $$;
+  my $newtask =
+    sub {
+      my $v = $task->();
+      kill 10 => $pid;
+      return $v
+    };
+  my $self = Async->new($newtask);
+  return unless $self;
+  $self->{callback} = $n; 
+  bless $self => AsyncCallback;
+}
+
+sub DESTROY {
+  my ($self) = @_;
+  my $n = $self->{callback};
+  undef $callbacks[$n];
+  my $i;
+  pop @callbacks until defined $callbacks[-1] || @callbacks == 0;
+  if (@callbacks == 0) {
+    $SIG{USR1} = $oldHandler;
+    undef $oldHandler;
+  }
 }
 
 1;
@@ -187,6 +316,9 @@ before C<$timeout> seconds have elapsed, it is forcibly terminated and
 returns a special value C<$special>.  The default special value is the
 string "Timed out\n".
 
+Because the timeouts are implemented with C<alarm()>, computations
+that use C<sleep()> or C<alarm()> will probably not work properly.
+
 All the other methods for C<AsyncTimeout> are exactly the same as for
 C<Async>.
 
@@ -198,7 +330,9 @@ C<Async>.
 C<AsyncData> is just like C<Async> except that instead of returning a
 string, the asynchronous computation may return any scalar value.  If
 the scalar value is a reference, the C<result()> method will yield a
-refernce to a copy of this data structure.
+reference to a copy of this data structure.  If the scalar value is
+I<not> a reference, the C<result()> method will I<still> yield a
+reference to the value.
 
 The C<AsyncData> module requires that C<Storable> be installed.
 C<AsyncData::new> will die if C<Storable> is unavailable.
@@ -207,6 +341,41 @@ All the other methods for C<AsyncData> are exactly the same as for
 C<Async>.
 
 
+=head1 TIED HASH
+
+Rather than managing the C<Async> objects directly, you can use
+C<Async> with a tied hash interface.  Use
+
+	tie %hash => Async;
+
+Then 
+
+	$hash{key} = sub { ... };
+
+runs the specified code asynchronously.  You can retrieve the result
+of the code by looking at the value of C<$hash{key}>.  If the code has
+not yet yielded a result, C<$hash{key}> will be undefined.  You have
+as many asynchronous jobs as you want; for example:
+
+	for $i (1 .. 100) {
+	  $hash{"square_root$i"} = sub { sqrt($i) };
+	}
+
+This starts 100 asynchronous jobs whose results will eventually appear
+in the hash.
+
+C<delete $hash{key}> destroys an asnychronous job without waiting for
+it to complete.  C<exists $hash{key}> returns true if and only if the
+job has completed.
+
+C<keys> will return a list of keys for computations that are running
+or have completed.  C<each> will return a key and the result of the
+corresponding asynchronous computation, or C<undef> if the compuation
+has not completed.  C<values> returns a list of results and C<undef>s.
+
+You can tie a hash to C<AsyncData> and get the data-passing semantics
+that it provides, but at present tying doesn't work for
+C<AsyncTImeout> because there's no way to pass the timeout.
 
 =head1 WARNINGS FOR THE PROGRAMMER
 
@@ -214,16 +383,17 @@ The asynchronous computation takes place in a separate process, so
 nothing it does can affect the main program.  For example, if it
 modifies global variables, changes the current directory, opens and
 closes filehandles, or calls C<die>, the parent process will be
-unaware of these things.  However, the asynchronous computatin does
+unaware of these things.  However, the asynchronous computation does
 inherit the main program's file handles, so if it reads data from
-files that the main program had open, that data will not be availble
+files that the main program had open, that data will not be available
 to the main program; similarly the asynchronous computation can write
 data to the same file as the main program if it inherits an open
 filehandle for that file.
 
 =head1 ERRORS
 
-The  errors that are reported by the C<error()> mechanism are: those that are internal to C<Async> itself:
+The only errors that are reported by the C<error()> mechanism are
+those that are internal to C<Async> itself:
 
 	Couldn't make pipe: (reason)
 	Couldn't fork: (reason)
@@ -234,6 +404,9 @@ considered to be an `error'; that is the normal termination of the
 process.  Any messages written to C<STDERR> will go to the
 computation's C<STDERR>, which is normally inherited from the main
 program, and the C<result()> will be the empty string.
+
+Compile-time errors in the code for the computation are, of course,
+caught at the time the main program is compiled.
 
 =head1 EXAMPLE
 
@@ -260,6 +433,7 @@ program, and the C<result()> will be the empty string.
 	print "The result of the computation is: ", $proc->result, "\n";
       }
       undef $proc;
+      last;
     }
     # The result is not ready; we can go off and do something else here.
     sleep 1; # One thing we could do is to take nap.
@@ -267,6 +441,6 @@ program, and the C<result()> will be the empty string.
 
 =head1 AUTHOR
 
-Mark-Jason Dominus C<mjd-perl-async+@plover.com>.
+Mark-Jason Dominus C<mjd-perl-async@plover.com>.
 
 =cut
